feat: lint and renaming things
