+++ server/bitbucketclient/client_server.go
@@ -1,4 +1,4 @@
-package bitbucket_server
+package bitbucketclient
 
 import (
 	"encoding/json"
