+++ server/bitbucketclient/factory.go
@@ -1,4 +1,4 @@
-package bitbucket_server
+package bitbucketclient
 
 import (
 	"fmt"
